Conflicts:
	examples/main.c
	src/cell.c
	src/common_io.h
	src/debug.c
	src/engine.c
	src/gravity/Default/gravity_io.h
	src/map.c
	src/single_io.c
	src/single_io.h
	src/space.c
	src/space.h

+++ src/single_io.h
@@ -26,8 +26,9 @@
 
 #if defined(HAVE_HDF5) && !defined(WITH_MPI)
 
-void read_ic_single(char* fileName, double dim[3], struct part** parts, struct gpart** gparts,
-		    int* Ngas, int* Ndm, int* periodic);
+void read_ic_single(char* fileName, double dim[3], struct part** parts,
+                    struct gpart** gparts, size_t* Ngas, size_t* Ndm,
+                    int* periodic);
 
 void write_output_single(struct engine* e, struct UnitSystem* us);
